Checkpoint - Deu ruim na mente, bugou geral, era pra ter começado com TDD, mas vacilei, agora vou fazer o TDD pra continuar o projeto, se não vou continuar bugado

+++ src/main/java/com/cabralbots/ecommerce/repository/ProductImageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
-    List<ProductImage> findAllByProduct_IdProduct(Integer idProduct);
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+    List<ProductImage> findAllByProduct_IdProduct(Long idProduct);
 }
