+++ src/main/java/com/cabralbots/ecommerce/repository/SessionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SessionRepository extends JpaRepository<Session, Integer> {
-    List<Session> findAllByUser_IdUserProduct_User(Integer userId);
-    Session findByUserProduct_IdUserProductAndUserIdUserProduct_User(Integer idProduct, Integer idUser);
+public interface SessionRepository extends JpaRepository<Session, Long> {
+    List<Session> findAllByUser_IdUserProduct_User(Long userId);
+    Session findByUserProduct_IdUserProductAndUserIdUserProduct_User(Long idProduct, Long idUser);
 }
